Rename p2p lib and modules
