+++ kaspad/Cargo.toml
@@ -14,7 +14,6 @@ consensus-core.workspace = true
 consensus.workspace = true
 rpc-core.workspace = true
 rpc-grpc.workspace = true
-kaspa-p2p.workspace = true
 p2p-flows.workspace = true
 
 thiserror.workspace = true
