+++ protocol/flows/Cargo.toml
@@ -9,7 +9,7 @@ edition = "2021"
 # project internal deps
 kaspa-core.workspace = true
 consensus-core.workspace = true
-kaspa-p2p.workspace = true
+p2p-lib.workspace = true
 # project external deps
 log.workspace = true
 futures = { version = "0.3", default-features = false, features = ["alloc"] }
