+++ protocol/p2p/src/bin/p2p_server.rs
@@ -2,11 +2,11 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 
 use kaspa_core::{debug, error};
-use kaspa_p2p_lib::kaspa_flows;
-use kaspa_p2p_lib::kaspa_flows::{EchoFlowRegistry, Flow};
-use kaspa_p2p_lib::kaspa_grpc;
-use kaspa_p2p_lib::kaspa_grpc::RouterApi;
-use kaspa_p2p_lib::kaspa_p2p::P2pAdaptorApi;
+use p2p_lib::adaptor::P2pAdaptorApi;
+use p2p_lib::infra;
+use p2p_lib::infra::RouterApi;
+use p2p_lib::registry;
+use p2p_lib::registry::{EchoFlowRegistry, Flow};
 
 #[tokio::main]
 async fn main() {
@@ -15,7 +15,7 @@ async fn main() {
     // [0] - init p2p-adaptor - server side
     let ip_port = String::from("[::1]:50051");
     let registry = Arc::new(EchoFlowRegistry::new());
-    let p2p_adaptor = kaspa_p2p_lib::kaspa_p2p::P2pAdaptor::listen(ip_port.clone(), registry).await.unwrap();
+    let p2p_adaptor = p2p_lib::adaptor::P2pAdaptor::listen(ip_port.clone(), registry).await.unwrap();
 
     let other_ip_port = String::from("http://[::1]:16111");
     for i in 0..1 {
@@ -34,20 +34,21 @@ async fn main() {
     tokio::time::sleep(std::time::Duration::from_secs(5)).await;
     debug!("P2P, p2p_server::main - FINISH");
 }
+
 #[allow(dead_code)]
 async fn old_main_with_impl_details() {
     // [-] - init logger
     kaspa_core::log::init_logger("trace");
     // [0] - Create new router - first instance
     // upper_layer_rx will be used to dispatch notifications about new-connections, both for client & server
-    let (router, mut upper_layer_rx) = kaspa_grpc::Router::new().await;
+    let (router, mut upper_layer_rx) = infra::Router::new().await;
     // [1] - Start service layer to listen when new connection is coming ( Server side )
     tokio::spawn(async move {
         // loop will exit when all sender channels will be dropped
         // --> when all routers will be dropped & grpc-service will be stopped
         while let Some(new_router) = upper_layer_rx.recv().await {
             // as en example subscribe to all message-types, in reality different flows will subscribe to different message-types
-            let (_flow_id, flow_terminate) = kaspa_flows::EchoFlow::new(new_router).await;
+            let (_flow_id, flow_terminate) = registry::EchoFlow::new(new_router).await;
             // sleep for 30 sec
             tokio::time::sleep(std::time::Duration::from_secs(30)).await;
             // terminate when needed (as an example) in general we need to save it somewhere in order to do graceful shutdown
@@ -55,7 +56,7 @@ async fn old_main_with_impl_details() {
         }
     });
     // [2] - Start listener (de-facto Server side )
-    let terminate_server = kaspa_grpc::P2pServer::listen(String::from("[::1]:50051"), router, true).await;
+    let terminate_server = infra::P2pServer::listen(String::from("[::1]:50051"), router, true).await;
     let terminate_signal = Arc::new(AtomicBool::new(false));
 
     // [3] - Check that server is ok & register termination signal ( as an example )
